Fix test compile issues
